requirements.txt
 PyPDF2==3.0.1
 pdfplumber==0.10.3
 python-docx==1.1.0
 pandas==2.2.0
 numpy==1.26.3
 scikit-learn==1.4.0
 streamlit==1.31.0
tabulate==0.9.0
